requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Async and HTTP client libraries
httpx==0.25.2
aiofiles==23.2.1

# Document parsing libraries
pypdf==3.17.1
python-docx==1.1.0

# Database libraries
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Vector database
pinecone-client==2.2.4

# LLM and embedding APIs
google-generativeai==0.3.2
openai==1.3.7  # For Jina embeddings API compatibility

# Utility libraries
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging libraries
python-json-logger==2.0.7

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
httpx==0.25.2  # For testing FastAPI endpoints
psutil==5.9.6  # For performance testing
reportlab==4.0.7  # For PDF generation in tests

# Development dependencies
black==23.11.0
flake8==6.1.0
mypy==1.7.1